apply models
