+++ packages/vlossom/src/components/vs-label-value/VsLabelValue.vue
@@ -18,7 +18,8 @@
 </template>
 <script lang="ts">
 import { PropType, computed, defineComponent, toRefs } from 'vue';
-import { getResponsiveProps, useColorScheme, useStyleSet } from '@/composables';
+import { useColorScheme, useStyleSet } from '@/composables';
+import { getResponsiveProps } from '@/models';
 import { VsComponent, type ColorScheme } from '@/declaration';
 import VsResponsive from '@/components/vs-responsive/VsResponsive.vue';
