+++ packages/vlossom/src/components/vs-tabs/VsTabs.vue
@@ -50,7 +50,8 @@
 
 <script lang="ts">
 import { computed, defineComponent, toRefs, ref, watch, onMounted, onUnmounted, type Ref, type PropType } from 'vue';
-import { useColorScheme, useStyleSet, getResponsiveProps, useIndexSelector } from '@/composables';
+import { useColorScheme, useStyleSet, useIndexSelector } from '@/composables';
+import { getResponsiveProps } from '@/models';
 import { VsComponent, type ColorScheme } from '@/declaration';
 import { VsIcon } from '@/icons';
 import { utils } from '@/utils';
