+++ packages/vlossom/src/composables/index.ts
@@ -9,6 +9,7 @@ export * from './input-modifier-composable';
 export * from './input-option-composable';
 export * from './layout-composable';
 export * from './modal-composable';
+export * from './overlay-composable';
 export * from './responsive-composable';
 export * from './scroll-lock-composable';
 export * from './state-class-composable';
