+++ packages/vlossom/src/models/index.ts
@@ -1,1 +1,2 @@
 export * from './input-model';
+export * from './responsive-model';
