+++ packages/vlossom/src/models/input-model.ts
@@ -1,6 +1,7 @@
-import { PropType } from 'vue';
 import { Message, Rule, UIState } from '@/declaration';
 import { utils } from '@/utils';
+
+import type { PropType } from 'vue';
 
 interface VsInputProps<T> {
     ariaLabel: { type: StringConstructor; default: null };
@@ -53,4 +54,12 @@ export function getInputProps<T = unknown, K extends keyof VsInputProps<T> = nev
         },
         excludes,
     );
+}
+
+export function getInputOptionProps() {
+    return {
+        options: { type: Array as PropType<any[]>, required: true, default: () => [] },
+        optionLabel: { type: String, default: '' },
+        optionValue: { type: String, default: '' },
+    };
 }
